database_api/api/requirements.txt
PyYAML == 5.3.1
SQLAlchemy == 1.3.19
SQLAlchemy_Utils == 0.36.8
cryptography == 2.8
fastapi == 0.61.1
hvac == 0.10.14
marshmallow == 3.13.0
marshmallow_sqlalchemy == 0.26.1
pydantic == 1.6.1
pytest == 6.0.1
python_jose == 3.2.0
requests == 2.22.0
starlette == 0.13.6
typer == 0.3.2
uvicorn == 0.11.8
python-dotenv==0.14.0
psycopg2-binary==2.8.6
